Convert compiler to return set_hl attrs sets, shuffle some files
